% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctemplates-package.R
\docType{package}
\name{doctemplates-package}
\alias{doctemplates}
\alias{doctemplates-package}
\title{doctemplates: Easy Document Templates for Grad Students}
\description{
This package provides R Markdown templates for common document types used grad students, like academic papers, presentations (forthcoming), and memos (forthcoming).
}
\author{
\strong{Maintainer}: Daniel Klein \email{daniel.klein@wisc.edu}

}
\keyword{internal}
